requirements.txt
requests>=2.5.0
openpyxl>=3.0.0
pysapgui>=1.0.7; sys_platform == 'win32'
PySapGUI~=1.0.7; sys_platform == 'win32'
setuptools~=57.0.0
PyYAML~=6.0.1